Initial commit for quota plugin
The quota plugin allows to enforce quotas in Gerrit.

To protect a Gerrit installation it makes sense to limit the resources
that a project or group can consume. To do this a Gerrit administrator
can use the quota plugin to define quotas on project namespaces.

With this change only a single quota for the maximum number of projects
in a namespace is supported. This quota is only of limited use for the
open source version of Gerrit since it doesn't allow to limit the
project creation capability to a namespace. The idea of this quota is
to allow a group A to create projects in the 'A/*' namespace and then
to limit the number of projects in that namespace. E.g. if this is a
paying customer to the number of projects that is defined in the
contract.

It is planned to add further quotas for maximum repository size and
maximum total size of repositories in a namespace in follow-up changes.

The defined quotas are stored in a `quota.config` file in the
`refs/meta/config` branch of the `All-Project` root project.

A namespace can be specified as
* exact project name (`plugins/myPlugin`): Defines a quota for one
  project.
* as patterns (`sandbox/*`): Defines a quota for one project
  namespace.
* as for-each-pattern (`public/?/*`): Defines the same quota for each
  subfolder.

If a project name matches several quota namespaces the one quota
applies to the project that is defined first in the `quota.config`
file.

Change-Id: I9b6ca2c94c3c7cdcf9b0bb0c6b91a760d8024505
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>