Add the maxTotalSize quota to limit total size of a namespace
For some scenarios it may be more convenient to only define quota on a
namespace level. In that case only the total size of all repositories
will be enforced.

This can also be combined with the maxRepoSize to enforce both the total
size of all repositories in a namespace and the max size of each of
them.

Change-Id: Iee40ea4740dbf64e03f14349caf32fac331d292d