Periodically publish fetch and push count
Count the numbers of pushes and fetches per project and publish
them to the registered UsageDataPublishedListeners.

After publishing, the counters are reset, so that the numbers of pushes
and fetches are counted for the time period since the last publishing
time.

Alternatives would have been to either accumulate forever or to publish
rates. The benefit of the chosen solution is that it is then easy for a
consumer which stores all data to accumulate sums over any interesting
time frame.

Depends-On: Ie1e82da47dc3906ef3da87dfbd0b583ebeecbef1
Change-Id: I5eb44390a650684030ebff2b12be6b30ba112a40