Let push and fetch counts survive server restart
The push and fetch counters were stored in a persistent cache that
was initialized with a loader with the value 0. The implementation
was flawed by the fact that the cached values are AtomicLongs that
were incremented in place. The cached instance never changed. Hence
the counters were never persisted to the SqlStore of the persistent
cache.

The cache has now been replaced by a ConcurentHashMap implementation.
To make the cached values survice (controlled) restart of ther server
the entries are now persisted to files in the plugins data directory
when the server is stopped. If the server is started, the counters
are initialized from the values in the files.

Change-Id: I0899e58ee7089862dcd05a8ce5fed3b7a9eaa2d5
Signed-off-by: Adrian Görler <adrian.goerler@sap.com>