Let push and fetch counts survive server restart
The cached values are of the AtomicLong type and they are updated
in-place. This meant that the cache was not aware of the updates
and did not mark the values as dirty. Therefore, they were not
persisted.

Perform a cache.put(key, atomicLong) on every update of the atomicLong
so that the value is internally marked as dirty.

Change-Id: I0899e58ee7089862dcd05a8ce5fed3b7a9eaa2d5