Update patch set 2
Patch Set 2:

> (1 comment)

This may fix the issue with the server restart (I haven't tested it either), but I think another motivation to not use a cache here is that the values cannot be recalculated if they are evicted from the cache. The solution proposed by Adrian would ensure that the values are not lost.

Patch-set: 2
Label: Verified=0