Update patch set 2
Patch Set 2:

> I don't think values would be lost with Saša's proposal. Data is
 > persisted continously. After server restart it would be loaded from
Yes, the values are persisted continuosly.
However, what happens if the cache decides to remove a value from cache (for whatever reason... maybe because some memory limit was reached)? Then the loader will not find
that value when it needs to be increased next time and will initialize a new one to zero.

 > the persistent cache. One could question though wether the data is
 > so valuable that it needs to be persisted continously. Or whether
 > it's cheaper to persist it upon server shutdown only.
If we can persist it continuously then we should do it. This can also be done asynchronously. 

So, I think we should first agree on data persistence (we cannot rely on the H2CacheImpl for that) and then see if we need the value in memory and, if yes, use a guava loading cache on top of the persistence model.
 > 
 > Also - why use the AtomicLong and not just a POL (plain old long).

Patch-set: 2