H2 based counter store
The fetch and push count data is predominantely written and rarely read.
Using a persistent cache is unsuitable for this use case.

Now the counters are stored using H2 in the plugin-specific data
directory. A cache might be added on top in a later change if the data
is read more frequently.

Change-Id: I0899e58ee7089862dcd05a8ce5fed3b7a9eaa2d5
Signed-off-by: Adrian Görler <adrian.goerler@sap.com>