Update patch set 1
Patch Set 1:

> where is 2.11-SNAPSHOT available, can't come from Maven central as
 > there only releases can be deployed

We can publish the SNAPSHOT artifacts in the Sonatype snapshot repository [1]. But it looks like snapshot artifacts get automatically deleted after a while... and we haven't deployed a 2.11 snapshot yet. At the moment this upload is done manually, if we need published snapshots, we must have a build job, which does this upload regularly.

[1] https://oss.sonatype.org/content/repositories/snapshots/com/google/gerrit/

Patch-set: 1
Label: Verified=0