Compute size by git-count-objects
Originally, the size of a repo is computed by the size of files on
disk. This patch set enables the option of computing repo size by
the output of "git count-objects -v" which reflects the disk
consumption of git-objects.

This option is useful when we want to avoid blocking end users
up to a dirty repo, e.g., GC does not run in time.

Here are some scenarios where we can use the new option, although
the answers to the questions at bottom are all negative.

1. repo folders are mounted to nfs volumes so that files in .nfs
   folders will be counted in quota.
2. bitmap files generated by git gc.
3. garbage files in the repo, captured by size-garbage entry. Those
   files ought to be deleted after garbage collection.

---Q&A---
Test with:
$ git --version
git version 1.9.1
$ jgit.sh --version
jgit version 4.1.2.201602141800-r

Does it consider only reachable objects?
No
1. clone a repo from server
2. git checkout SH1 of second last commit. Detached head
3. commit a new file (30mb, for simplicity)
4. git checkout master
5. git reflog expire --expire=now --all
6. git fsck --full (this will show a dangling commit)
7. git count-objects to show the size: 30 is in loose object

If an object is present in more than one pack file, is its size
computed only once?

No
Steps
1. clone a repo from server
2. git config gc.pruneExpire "2.days.ago"
3. commit a new file1 (30mb, for simplicity)
4. jgit gc
5. git count-objects to show the size: 30 is in pack1
6. commit a new file2 (20mb)
7. jgit gc
8. git count-objects to show the size: 80 where 30 in pack1 (only the
   first commit) and 50 in pack2 (2 commits)

If an object is present in a pack file and also as a loose object, is
its size computed only once?

No
Steps:
1. clone a repo from server
2. commit a new file (30mb, for simplicity)
3. back up the object file containing that commit object
4. git count-objects to show the size: 30 is in loose object size as
   expected
5. jgit gc
6. git count-objects to show the size: 30 is in pack size as expected
7. recover the object file. Now the object is in both pack and loose
   object
8. git count-objects to show the size: 30 is in both pack size and
   loose object size

Change-Id: Icc5bc6fdf8fa5c4635eabc1cd61870890e5277c9