Compute size by git-count-objects
Originally, the size of a repo is computed by the size of files on
disk. This patch set enables the option of computing repo size by
the output of "git count-objects -v" which is an estimation of the
repo size if we run garbage collection on the repo.

This option is useful when we want to avoid blocking end users
up to a dirty repo, e.g., GC does not run in time.

Change-Id: Icc5bc6fdf8fa5c4635eabc1cd61870890e5277c9