Update patch set 1
Patch Set 1:

> Why is this done for stable-2.12?  The buck change that removed
 > source_under_test is only on gerrit's master branch.
 > 
 > https://gerrit-review.googlesource.com/#/c/84030/

The problem is that latest bucklets version is using that buck version. Having the bucklets as a submodule to pin a version or having stable branches in the bucklets would avoid needing this change.

This change is backward compatible because it would also work with the old buck version except for the fact that code coverage report would not be available.

Patch-set: 1