Create a provider to retrieve quota.config
Encapsulate how to retrieve quota.config in a guice provider class.
This refactoring will allow to support another location for quota.config
without having to duplicate code to find it.

Currently, quota.config must be in refs/meta/config branch of
All-Projects but a follow up commit will add the possibility to put it
in review_site/etc directory.

Both classes using quota config are singleton so inject the provider of
config instead of the config directly to make sure previous behavior is
preserved, which is to get the lastest version of quota.config every
time it is needed.

Change-Id: If2f428d7694b4f0389f43a46b023c4210c574831