Support quota.config in review_site/etc folder
Storing the quota.config in All-Projects prevents having different rates
on master and slave because All-Projects on slave is a replica of the
master one.

To support this use case, quota.config can now be stored in
review_site/etc folder. If quota.config is found in both All-Projects
and review_site/etc, plugin will fail to load with a proper error
message.

Change-Id: I9080cda8aa17691ab32c5c8889ff8707a3df0034