Add configuration option to disable disk quota
Currently, this plugin is comprised of the rate limit feature, as well
as the disk quota feature. In cases where disk quota does not need to
be enforced, the plugin will still cache and track repo sizes. This
introduces unneeded overhead. This change adds a configuration option to
disable the disk quota cache.

Change-Id: Ic3c4bbdca3a7aa195e84e3ff40f86102ade1a443