Allow burst value to be optional in the configuration
So far, it was necessary to pass the burst parameter to the
configuration for it to be recognized as valid even if the default
value of 30 was appropriate.

Modify the regex so that the burst is now optional. In this case, the
configuration:

  [group "<groupName>"]
    <rateLimitType> = <rateLimit> <rateUnit>

is recognized as valid and the burst is set to the default value.

Change-Id: Ib5a6173667a50e3435e1863a7a075bc9043e3479