Update patch set 4
Patch Set 4:

> > > @Marco do you have any idea how i can get this to work?
> > 
> > I think this review (previously WIP) now became obsolete, as per the new "blocks" plugin proposal [1].
> > 
> > [1] https://groups.google.com/forum/#!topic/repo-discuss/naYbybi3E2w
> 
> What's the new blocks proposal?

Yours, no? -Cf. [1] right above. My understanding is that this [1] (the thread) ends with replies about link(s) pointing to your newly proposed plugin, called "blocks" or so. This would mean that we no longer need this very change here in quota.

If otherwise, then a minimal design doc might be needed, so we understand how quota and blocks plugins need to be integrated, for this topic to work as a whole.

Patch-set: 4
Reviewer: Gerrit User 1012732 <1012732@173816e5-2b9a-37c3-8a2e-48639d4f1153>