Add namespacing support for task quotas
Make it possible to specify quotas for a project namespace.
It utilizes the already existing syntax for `quota` section.

The project name is correctly decoded when the specific task
has it specified it in its `toString()` implementation. If a
project name was not found quotas from the global (`*`) section
are applied.

Change-Id: I81ede1412c5876b9f89f73e42557fc56532aa464