Only have a single shared SoftMax quota
```
[quota "project-a"]
    softMaxStartPerUserForQueue = 5 SSH-Interactive-Worker
[quota "project-b"]
    softMaxStartPerUserForQueue = 10 SSH-Interactive-Worker
```

Previously, for the above configurations, two seprate TaskQuota
instances were created in the background and hence computation
of free threads made on one quota is not reflected on the other.
Fix this to only have single soft max quota per queue which has
details of all the tasks that are being ran in the queue.

Change-Id: Ib8e82fc413683647b9ef730819121bcb2d42b5c2