Only have a single shared free threads/permits for soft max
```
[quota "project-a"]
    softMaxStartPerUserForQueue = 5 SSH-Interactive-Worker
[quota "project-b"]
    softMaxStartPerUserForQueue = 10 SSH-Interactive-Worker
```

Previously, for the above configuration, two separate TaskQuota (free
threads) instances were created in the background. The computation of
free threads was buggy since its not shared. Each instance is not aware
of the permits acquired by the other quota. This results in config not
being enforced correctly. Fix this to only have single soft max quota
per queue which has details of all the tasks that are being ran in the
queue.

Change-Id: Ib8e82fc413683647b9ef730819121bcb2d42b5c2