Fix MaxRepositorySizeQuota.dryRun method
When change Ibb4d282ebb introduced the `dryRun` method to the
MaxRepositorySizeQuota, it erroneously implemented the `requestQuota`
method so that it would always return an error response, even when the
requesting available tokens.

The lack of testing prevented this issue from being detected, so this
change also introduces a test coverage for the
`MaxRepositorySizeQuota.dryRun` method, to ensure expected behaviour and
prevent future unnoticed failures.

Change-Id: I949fe894fa719b30d0632f542829916f4dc8c726