Fix task quotas not being released
In I1d2949d84, assumption was made that TaskParker#onNotReadyToStart is
called by WorkQueue when TaskParker#isReadyToStart returns false and
removed the cleaning up of acquired quotas. But that only happens when
different a TaskParker instance returns false but not the current one.
Fix that assumption by re-introducing the explicit cleanup logic.

Change-Id: Iea32b57b86acce0a96d8f29789eba3fcee234d99