Update repo_size persistent cache
The repo_size cache (introduced by Ia6be19c7f40), is updated when a
receive-pack is successfully executed, so that the repository size can
be updated accordingly rather than having to recompute the repository
size, which is an expensive operation.

The previous code however only updated the **values** of the in-memory
cache, without actually updating the cache itself, via the `put` method.

As a consequence, changes to the repo_size cache were never persisted to
disk and they were lost when the Gerrit process was restarted.

Ensure the cache is properly updated, so that both the in-memory and the
persistent cache layer are updated.

Bug: Issue 464448637
Change-Id: I2652482c3aa43dd34ec59cdb763ad8757b40d3ea